// packages/ui-lib/src/theme/colors.module.scss.d.ts
export interface AppColors {
    blue: '#2b0099';
    darkBlue: '#210075';
    lightBlue: '#dfd9f0';
    ultraLightBlue: '#f4f2fa';

    green: '#00d7a0';
    lightGreen: '#def9f2';
    ultraLightGreen: '#f2fdfa';

    white: '#ffffff';
    ultraLightGray: '#fafafa';
    lightGray: '#eeeeee';
    strokeGray: '#dddddd';
    mediumGray: '#cccccc';
    warmGray: '#999999';
    darkGray: '#646464';
    black: '#000000';

    yellow: '#ffbb00';
    lightYellow: '#fffbef';

    darkRed: '#9e0738';
    red: '#f65a5a';
    lightRed: '#fff6f6';

    pink: '#ffc7c7';
    darkPink: '#cf9898';
    skyBlue: '#739fff';
    golden: '#dcb382';
    grayBlue: '#627d9b';
    turquoise: '#61e3f1';
    // old colors
    primaryDark: '#3c00b7';
}

declare const colors: AppColors;
export default colors;